Move Predicates class to util package
