+++ presto-main/src/main/java/com/facebook/presto/util/Predicates.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.facebook.presto.sql.planner.optimizations;
+package com.facebook.presto.util;
 
 import java.util.function.Predicate;
